Support "Open in Object Browser" in dependencies tree
This commit adds support for the context menu item "Open in Object Browser", which is added to context menus in the dependencies tree.

The previous command handler only suppressed the menu item in nearly all cases.

This new code adds support for nodes to advertise their support for this feature, and to provide a path to the DLL to browse.

New groups/items are added to the relevant context menus in the VSCT file.

A follow-up change will be made to NuGet, to allow browsing to package DLLs using the new `IObjectBrowserItem` interface added here.