Fix NFE about version regression on configuration switch
In #8321, we optimised how we handle Roslyn projects in the language service during configuration switches, such as from Debug to Release.

Before that PR, we maintained different pipelines for each configuration. That PR re-uses the same pipeline between configurations, instead applying a delta representing the switch. This makes configuration switches a lot faster.

Project data flows through with version numbers, and each configuration tracks versions to ensure that evaluation and build data is applied with correct sequencing.

The previous PR introduced the chance for an unexpected regression in configured project version, which would lead to an exception that'd trigger a NFE and gold bar in VS.

> Attempted to push a project evaluation that regressed in version.

This is one of the top NFEs in the project system.

The fix here is to consider multiple versions in each update. We combine a version that increments when the active configuration changes along with the actual configured project version. Then, when comparing versions, we first compare the active configuration version, and only if they match would we then compare the configured project version. This gives us a monotonically increasing sequence of updates across both configuration switches and project changes, which will satisfy the consuming code and prevent the NFE.