Close info bars when associated project(s) unload
Previously, info bars would remain in VS after solution close, or after project unload. This can lead to confusion when opening a new solution and seeing an error message from the previously open solution.

This change keeps track of which project(s) are associated with each info bar, and monitors the unload event of those projects. Now, when a project unloads, it will remove the info bar if no other remaining projects are also associated with that info bar.