Add .git-blame-ignore-revs
This file tells git/GitHub to ignore certain commits from the blame view.

The file is configured to exclude a recent solution-wide conversion from block-scoped to file-scoped namespaces, which touched most lines in the repo. With this, GitHub's blame view will ignore the change, making browsing changes on GitHub easier.