Fix "collection modified" exception
When signalling project close to an info bar entry, if there are no other remaining projects associated with that info bar, it will be removed from the collection. That triggers an exception.

The fix is to make a defensive copy before enumerating, so that the original can be safely modified while the copy is enumerated.