FUTDC finds source items in both File and Project schema
In https://devdiv.visualstudio.com/DevDiv/_git/CPS/pullrequest/587251, CPS made a change to deduplicate items from different rules mapping to the same item type. This inadvertently broke incremental builds for WPF projects, where changes to `.xaml` files (`Page` items) would not cause the project to be considered as needing to build (underbuild) which is very frustrating and confusing for users.

During the investigation of this issue it was discovered the items returned by CPS's `SourceItemsRuleSource` come from both the `File` and `Project` catalogs. Historically the FUTDC has only checked the `File` catalog for these items. When CPS introduced deduplication in that PR, the `Page` rule for WPF projects became present in only the `Project` catalog.

This change teaches the FUTDC to check in both the `File` _and_ `Project` catalogs. This fixes the WPF issue, and potentially others, by covering all origins for source items that CPS reports.