Inline method and fix NullReferenceException
The `IVsDebuggerLaunchCompletionCallback` interface is not null-annotated, and the last argument is not treated as nullable by consuming code. We were passing null for that argument in some cases.

By inlining the method and using our concrete type, we have an annotated type that highlights the bug.

We also avoid a redundant collection copy to produce an array.