Treat warnings as errors in release builds
Lately we've started seeing warnings creeping into the codebase. These should be caught by CI and prevented.

To make it easier to reproduce CI behaviour locally, this setting is enabled for release builds. If CI fails and you want to recreate the error locally, switch to release mode and build.