Remove VSSDK FUTDC overrides
These targets could cause issues during DTBs (The target "GetVsixSourceItems" does not exist in the project.) if the VSSDK wasn't restored.

Newer versions of the VSSDK implement the equivalent logic.

This change removes our explicit overrides. If we see any issues with the FUTDC after this we can work with the VSSDK team to address any gaps.