Disallow properties in rules to start with underscore
The project data caches don't include any properties that start with underscores, as these are considered 'private' and are therefore ignored to save cache space. Having a rule that uses a private property means that the consumer will see different data depending upon whether it comes from MSBuild or from a cache. The difference in behaviour would very likely cause downstream problems, so we should disallow it.