Integrate into new Visual Studio retargeting UI (#9771)
This change integrates into the new retargeting UI in Visual Studio 2026.

When a global.json is present and it is not the latest version in that feature band we will supply retargeting information to Visual Studio to allow the customer to visit the download link to install a suggested SDK.

At this time the suggestion is confined to the same Major.Minor sdk which was specified in the global.json. Note that we will make this suggest even if the customer has a later SDK version globally installed.

We do not, as of this version update global.json or initiate the download or install.